refactor: entity별 폴더구조 변경
