+++ src/basic/components/Header.tsx
@@ -1,7 +1,6 @@
-import { CartItemType } from '../../types';
-import { SearchInput } from '../ui/SearchInput';
+import { CartItemType } from '@/types';
+import { SearchInput, Button } from '../ui';
 import { CartIcon } from '../icons';
-import { Button } from '../ui/Button';
 
 interface HeaderPropsType {
   isAdmin: boolean;
