--- src/basic/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './product';
-export * from './coupon';
