+++ src/basic/entities/cart/hooks/useCart.ts
@@ -1,7 +1,7 @@
 import { Dispatch, useCallback, useEffect, useState } from 'react';
-import { CartItemType } from '../../types';
-import { ProductWithUI } from '../types/product';
-import { getRemainingStock } from '../utils';
+import { CartItemType } from '@/types';
+import { ProductWithUI } from '../../product';
+import { getRemainingStock } from '../../../utils';
 
 interface UseCartPropsType {
   products: ProductWithUI[];
