+++ src/basic/entities/coupon/index.ts
@@ -1,0 +1,13 @@
+// Components
+export { Coupon } from './components/Coupon';
+export { AdminCoupon } from './components/AdminCoupon';
+
+// Hooks
+export { useCoupon } from './hooks/useCoupon';
+export { useCouponForm } from './hooks/useCouponForm';
+
+// Types
+export * from './types/coupon';
+
+// Constants
+export * from './constants/coupon';
