+++ src/basic/entities/payment/index.ts
@@ -1,0 +1,5 @@
+// Components
+export { Payment } from './components/Payment';
+
+// Hooks
+export { usePayment } from './hooks/usePayment';
