+++ src/basic/entities/product/constants/product.ts
