+++ src/basic/entities/product/hooks/useProductSearch.ts
