+++ src/basic/entities/product/index.ts
@@ -1,0 +1,14 @@
+// Components
+export { Product } from './components/Product';
+export { AdminProduct } from './components/AdminProduct';
+
+// Hooks
+export { useProduct } from './hooks/useProduct';
+export { useProductForm } from './hooks/useProductForm';
+export { useProductSearch } from './hooks/useProductSearch';
+
+// Types
+export * from './types/product';
+
+// Constants
+export * from './constants/product';
