+++ src/basic/hooks/index.ts
@@ -1,7 +1,3 @@
 export * from './useDebounce';
 export * from './useNotification';
 export * from './useLocalStorage';
-export * from './useProductSearch';
-export * from './useCart';
-export * from './useCoupon';
-export * from './usePayment';
