+++ src/basic/pages/index.ts
@@ -1,0 +1,2 @@
+export { MainPage } from './MainPage';
+export { AdminPage } from './AdminPage';
